Phase 8: Secondary screens + Milestone 2 report
- TracksScreen with TrackMixerRow (mute/solo/quantize/loop/volume/instrument/delete)
- TrackFocusScreen with PianoRollCanvasView + DrumGridView
- SettingsScreen + AboutScreen
- All test tags matching iOS accessibility identifiers
- assembleDebug PASS, all 187 tests PASS
- Milestone 2 report covering Phases 5-8

+++ android-app/feature/editor/src/main/java/com/loopa/feature/editor/ui/PianoRollCanvasView.kt
@@ -51,7 +51,7 @@ fun PianoRollCanvasView(
                     if (tappedNote != null) {
                         onNoteTap(tappedNote.id)
                     } else {
-                        onEmptyTap(tappedBeat, tappedPitch.coerceIn(0, 127).toUByte())
+                        onEmptyTap(tappedBeat.toDouble(), tappedPitch.coerceIn(0, 127).toUByte())
                     }
                 }
             }
